backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
psycopg[binary]==3.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
scikit-learn==1.4.0
pandas==2.2.0
openpyxl==3.1.2
httpx==0.26.0
pytest==0.0
pytest-asyncio==0.23.5
python-dotenv==1.0.1
numpy==1.26.4
supabase==2.16.0
